narrow: Extract narrow_title module.
