+++ web/src/views_util.js
@@ -3,6 +3,7 @@ import $ from "jquery";
 import * as message_view_header from "./message_view_header";
 import * as narrow from "./narrow";
 import * as narrow_state from "./narrow_state";
+import * as narrow_title from "./narrow_title";
 import * as navigate from "./navigate";
 import * as pm_list from "./pm_list";
 import * as resize from "./resize";
@@ -35,7 +36,7 @@ export function show(opts) {
     unread_ui.hide_unread_banner();
     opts.update_compose();
     narrow_state.reset_current_filter();
-    narrow.update_narrow_title(narrow_state.filter());
+    narrow_title.update_narrow_title(narrow_state.filter());
     message_view_header.render_title_area();
     narrow.handle_middle_pane_transition();
     search.clear_search_form();
